+++ src/app/components/content-clusters-nodes.js
@@ -3,7 +3,7 @@
  *
  * @module components/content-cluster-nodes
  * @author Jakub Liput
- * @copyright (C) 2017 ACK CYFRONET AGH
+ * @copyright (C) 2017-2018 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
@@ -11,13 +11,31 @@ import Component from '@ember/component';
 
 import { inject as service } from '@ember/service';
 import { Promise } from 'rsvp';
+import { computed } from '@ember/object';
+import I18n from 'onedata-gui-common/mixins/components/i18n';
 
 import PromiseObject from 'onedata-gui-common/utils/ember/promise-object';
+import clusterIpsConfigurator from 'onepanel-gui/mixins/components/cluster-ips-configurator';
+import safeExec from 'onedata-gui-common/utils/safe-method-execution';
+import _ from 'lodash';
 
-export default Component.extend({
+export default Component.extend(I18n, clusterIpsConfigurator, {
   onepanelServer: service(),
   clusterManager: service(),
+  providerManager: service(),
   globalNotify: service(),
+
+  i18nPrefix: 'components.contentClustersNodes',
+
+  /**
+   * @override
+   * @type {PromiseObject<ProviderDetails>}
+   */
+  providerDetailsProxy: computed(function getProviderDetailsProxy() {
+    if (this.get('onepanelServer.serviceType') === 'provider') {
+      return this.get('providerManager').getProviderDetails();
+    }
+  }),
 
   /**
    * Resolves with EmberArray of ClusterHostInfo.
@@ -30,6 +48,16 @@ export default Component.extend({
    * @type {string}
    */
   primaryClusterManager: null,
+
+  /**
+   * If true, table of IPs is in edition mode
+   * @type {boolean}
+   */
+  ipsEdit: false,
+
+  _ipsFormData: undefined,
+
+  _origIpsFormData: undefined,
 
   init() {
     this._super(...arguments);
@@ -57,5 +85,31 @@ export default Component.extend({
         promise: clusterHostsPromise,
       })
     );
+  },
+
+  /**
+   * @override
+   */
+  _startSetup() {
+    return this._super(...arguments).then(() => {
+      safeExec(this, 'set', 'ipsEdit', false);
+    });
+  },
+
+  actions: {
+    enableIpsEdit() {
+      this.setProperties({
+        ipsEdit: true,
+        _origIpsFormData: _.cloneDeep(this.get('_ipsFormData')),
+      });
+    },
+
+    cancelIpsEdit() {
+      this.setProperties({
+        ipsEdit: false,
+        _ipsFormData: this.get('_origIpsFormData'),
+        _origIpsFormData: undefined,
+      });
+    },
   },
 });