+++ src/app/locales/en/components/cluster-host-ip-form.js
@@ -1,0 +1,7 @@
+export default {
+  hostname: 'Hostname',
+  ip: 'IP address',
+  hostRow: {
+    example: 'Example: 172.17.0.2',
+  },
+};