+++ src/app/locales/en/components/content-clusters-nodes.js
@@ -1,0 +1,8 @@
+export default {
+  header: 'Nodes of "{{name}}"',
+  tabServices: 'Services',
+  tabIps: 'Cluster IPs',
+  editIps: 'Edit IPs',
+  cancelIpsEdit: 'Cancel editing',
+  setupIpsButton: 'Confirm changes',
+};