+++ src/app/locales/en/components/new-cluster-ips.js
@@ -1,0 +1,8 @@
+export default {
+  header: 'IP adresses setup',
+  setupIpsButton: 'Setup IP adresses',
+
+  infoAutoDetect: 'Below IP adresses are auto-detected suggestions that need to ' +
+    'be revised. Please make sure that your cluster nodes are reachable at ' +
+    'selected IP addresses or modify them accordingly.',
+};