+++ src/app/locales/en/components/new-cluster.js
@@ -2,13 +2,15 @@ export default {
   steps: {
     provider: {
       installation: 'cluster installation',
-      providerRegistration: 'zone registration',
+      ips: 'cluster IP adresses',
+      providerRegistration: 'provider registration',
       providerCert: 'certificate setup',
       providerStorage: 'storage configuration',
       summary: 'summary',
     },
     zone: {
       installation: 'cluster installation',
+      ips: 'cluster IP adresses',
       summary: 'summary',
     },
   },